Updated to work with Spring Boot 3 and Jakarta EE
Updated the library to work with Spring Boot 3, which uses Jakarta EE
instead of Java EE 8.

This caused several imports to change, making this upgrade a breaking
change which is incompatible with Spring Boot 2.x.

Also updated various dependencies, and made lombok a provided dependency
as it should have been.